Added LLM call to Flask API. Reformated imports to be relative and split supporting functions into utils.py
